fix: qos assocs does not imply available cluster
The rationale for calculating a denied list rather than an allowed list
is that one cannot violate either of the conditions (not in allowlist
or in disallow list) allows it to be a union operation rather than an
intersection operation that requires additional maintenance of number
of presences in list. Considering that the account x qos could be pretty
small, it might be better to just keep the code a nice look.